skill parser: extract CharacterSkills::EffectParser
Description -> ordered effect rows (clause-based targets, status matching,
damage/heal/dispel) moves out, owning the unmatched-status accumulator and the
TT/CONTEXT/DEFAULT_TARGETS/STACKING constants. Parser injects one instance.
